pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Софья Григорович <sonyashkin@icloud.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
